+++ lib/src/controller/store.dart
@@ -1,5 +1,6 @@
 import '../core/nodes.dart';
 import '../core/scene.dart';
+import '../core/id_generator.dart';
 import '../model/document.dart';
 import '../model/document_clone.dart';
 
@@ -10,14 +11,14 @@ class SceneStore {
           : Set<NodeId>.from(selectedNodeIds),
       allNodeIds = txnCollectNodeIds(sceneDoc),
       nodeLocator = txnBuildNodeLocator(sceneDoc),
-      nodeIdSeed = txnInitialNodeIdSeed(sceneDoc),
-      layerIdSeed = txnInitialLayerIdSeed(sceneDoc),
+      idGeneratorState = createInitialIdGeneratorState(sceneDoc),
       nextInstanceRevision = txnInitialNodeInstanceRevisionSeed(sceneDoc);
 
   Scene sceneDoc;
   Set<NodeId> selectedNodeIds;
   Set<NodeId> allNodeIds;
   Map<NodeId, NodeLocatorEntry> nodeLocator;
+  IdGeneratorState idGeneratorState;
 
   int controllerEpoch = 0;
   int structuralRevision = 0;
@@ -25,7 +26,15 @@ class SceneStore {
   int visualRevision = 0;
   int commitRevision = 0;
 
-  int nodeIdSeed;
-  int layerIdSeed;
   int nextInstanceRevision;
+
+  int get nodeIdSeed => idGeneratorState.nextNodeCounter;
+  set nodeIdSeed(int value) {
+    idGeneratorState.nextNodeCounter = value;
+  }
+
+  int get layerIdSeed => idGeneratorState.nextLayerCounter;
+  set layerIdSeed(int value) {
+    idGeneratorState.nextLayerCounter = value;
+  }
 }